Build with LEFTOVERDEBUG and LEFTOVERSPECTRUM like the US ROM
gepc-ref's ntsc-final and the N64 US Makefile set both. Reviewed every block
first: US-ROM gameplay values (watch perspective, aim timing, stan edge test
variant), print-only model checks, and debug/Spectrum menus. Provide
get_counters() for the speed graph, fix two debug-only pointer casts, and
regenerate the implicit prototypes for the newly compiled callers.

+++ src/debugmenu.c
@@ -356,7 +356,7 @@ Gfx *debmenuDraw(Gfx *gdl)
 
 	// Make sure there'll be a least 256 GBI commands free (2KB)
 	available = dynGetFreeGfx(gdl) - 256 * sizeof(Gfx);
-	needed = (u32)gdl2 - (u32)gdl;
+	needed = (u32)((uintptr_t)gdl2 - (uintptr_t)gdl);
 
 	if (needed <= 0) { // shouldn't be possible
 		return gdl;
